Add get order service

+++ app/featues/order/form/order_item.go
@@ -3,6 +3,6 @@ package form
 type OrderItem struct {
 	ProductId string  `json:"productId" binding:"required"`
 	Quantity  int     `json:"quantity" binding:"required"`
-	Price     float32 `json:"price" binding:"required"`
-	Discount  float32 `json:"discount"`
+	Price     float64 `json:"price" binding:"required"`
+	Discount  float64 `json:"discount"`
 }
